Move notify component to bigcommerce folder
